+++ .npmignore
@@ -1,0 +1,7 @@
+dist
+src/stories
+.gitignore
+.prettierrc
+.eslintrc.json
+rollup.config.js
+tsconfig.json
