+++ package.json
@@ -1,11 +1,14 @@
 {
+    "name": "template-editor",
+    "description": "Mail template editor module",
+    "version": "0.1.0",
     "scripts": {
         "dev": "storybook dev",
-        "dev-vite": "vite --host",
-        "build": "tsc && vite build",
-        "preview": "vite preview"
+        "build": "rollup -c"
     },
+    "type": "module",
     "devDependencies": {
+        "@rollup/plugin-json": "^6.0.0",
         "@storybook/addon-essentials": "^7.0.12",
         "@storybook/addon-interactions": "^7.0.12",
         "@storybook/addon-links": "^7.0.12",
@@ -25,6 +28,7 @@
         "eslint-plugin-react-refresh": "^0.4.1",
         "eslint-plugin-storybook": "^0.6.12",
         "prop-types": "^15.8.1",
+        "rollup-plugin-typescript2": "^0.34.1",
         "storybook": "^7.0.12",
         "typescript": "^5.0.4",
         "vite": "^4.3.6",
@@ -33,5 +37,12 @@
     "dependencies": {
         "react": "^18.2.0",
         "react-dom": "^18.2.0"
-    }
+    },
+    "peerDependencies": {
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0"
+    },
+    "files": [
+        "dist"
+    ]
 }
