+++ src/components/TestComponent/TestComponent.tsx
@@ -1,0 +1,3 @@
+export const TestComponent = () => {
+    return <div>TestComponent</div>;
+};
