+++ src/components/TestComponent/index.ts
@@ -1,0 +1,1 @@
+export { TestComponent as default } from './TestComponent';
