+++ src/stories/Readme.mdx
@@ -1,0 +1,11 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Readme/Stuff" />
+
+<style></style>
+
+# Some Markdown content
+
+Some text
+
+{/* HTML IS ALLOWED TOO */}
